+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+**9.1.0**
+
+```
+IMP: file utils operations
+
+FIX: TrasSH: 'inline' load runner result data attribute (issue #92)
+FIX: Init. Commands (issue #88)
+
+ADD: Command 'genfile': Generates a File object
+ADD: Math operations
+```
+
 **9.0.0**
 
 ```